chore: move `openai` module to `analyser` module
