+++ backend/src/module/parser/analyser/__init__.py
@@ -2,3 +2,4 @@ from .raw_parser import raw_parser
 from .tmdb_parser import tmdb_parser
 from .torrent_parser import torrent_parser
 from .mikan_parser import mikan_parser
+from .openai import OpenAIParser
