+++ backend/src/module/parser/analyser/openai.py
